# openapi_patched.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "默认模块",
    "description": "",
    "version": "1.0.0"
  },
...
  "paths": {
...
    "/api/v1/orders": {
      "post": {
        "summary": "创建订单",
        "deprecated": false,
        "description": "【版本】v1\n【环境】REST 接口\n【前置条件】需要用户登录\n【功能描述】根据购物车选择的商品或直接购买创建新的订单",
        "operationId": "创建订单",
        "tags": [
          "订单管理"
        ],
        "parameters": [],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateOrdersRequest"
              },
              "example": {
                "addressId": 501,
                "items": [
                  {
                    "productId": 2005,
                    "quantity": 1
                  }
                ],
                "note": "周末派送"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "成功",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CreateOrdersResponse"
                },
                "example": {
                  "code": 200,
                  "data": {
                    "orderId": 10001,
                    "orderNo": "ORD20260411123456"
                  },
                  "message": "订单创建成功"
                }
              }
            },
            "headers": {},
            "x-apifox-name": "成功",
            "x-apifox-ordering": 0
          },
          "400": {
            "description": "请求参数错误",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                },
                "example": {
                  "code": 400,
                  "message": "请求参数错误",
                  "details": {
                    "field": "name",
                    "reason": "不能为空"
                  }
                }
              }
            },
            "headers": {},
            "x-apifox-name": "请求有误",
            "x-apifox-ordering": 1
          },
          "401": {
            "description": "未授权",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                },
                "example": {
                  "code": 401,
                  "message": "未授权，请先登录"
                }
              }
            },
            "headers": {},
            "x-apifox-name": "没有权限",
            "x-apifox-ordering": 2
          },
          "403": {
            "description": "禁止访问",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                },
                "example": {
                  "code": 403,
                  "message": "无权限访问此资源"
                }
              }
            },
            "headers": {},
            "x-apifox-name": "禁止访问",
            "x-apifox-ordering": 3
          },
          "404": {
            "description": "资源不存在",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                },
                "example": {
                  "code": 404,
                  "message": "请求的资源不存在"
                }
              }
            },
            "headers": {},
            "x-apifox-name": "记录不存在",
            "x-apifox-ordering": 4
          },
          "409": {
            "description": "资源冲突",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                },
                "example": {
                  "code": 409,
                  "message": "资源已存在或状态冲突"
                }
              }
            },
            "headers": {},
            "x-apifox-ordering": 5
          },
          "422": {
            "description": "实体无法处理",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                },
                "example": {
                  "code": 422,
                  "message": "请求格式正确但语义错误",
                  "details": {
                    "field": "email",
                    "reason": "格式不正确"
                  }
                }
              }
            },
            "headers": {},
            "x-apifox-name": "参数错误",
            "x-apifox-ordering": 6
          },
          "500": {
            "description": "服务器内部错误",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                },
                "example": {
                  "code": 500,
                  "message": "服务器内部错误，请稍后重试"
                }
              }
            },
            "headers": {},
            "x-apifox-name": "服务器错误",
            "x-apifox-ordering": 7
          },
          "502": {
            "description": "网关错误",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                },
                "example": {
                  "code": 502,
                  "message": "网关错误，上游服务不可用"
                }
              }
            },
            "headers": {},
            "x-apifox-name": "网关错误",
            "x-apifox-ordering": 8
          },
          "503": {
            "description": "服务不可用",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                },
                "example": {
                  "code": 503,
                  "message": "服务暂时不可用，请稍后重试"
                }
              }
            },
            "headers": {},
            "x-apifox-name": "服务器故障",
            "x-apifox-ordering": 9
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "x-apifox-folder": "订单模块",
        "x-apifox-status": "released",
        "x-run-in-apifox": "https://app.apifox.com/web/project/8090070/apis/api-441880932-run"
      },
      "get": {
        "summary": "分页查询订单列表",
        "deprecated": false,
        "description": "【版本】v1\n【环境】REST 接口\n【前置条件】需要用户登录\n【功能描述】分页查询当前用户的订单列表，支持按状态筛选",
        "operationId": "分页查询订单列表",
        "tags": [
          "订单管理"
        ],
        "parameters": [
          {
            "name": "page",
            "in": "query",
            "description": "当前页码，默认 1",
            "required": false,
            "schema": {
              "type": "integer"
            }
          },
          {
            "name": "pageSize",
            "in": "query",
            "description": "每页记录数，默认 10",
            "required": false,
            "schema": {
              "type": "integer"
            }
          },
          {
            "name": "status",
            "in": "query",
            "description": "订单状态（例如: pending, paid, shipped, completed, cancelled）",
            "required": false,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "成功",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OrdersListResponse"
                },
                "example": {
                  "code": 200,
                  "data": {
                    "items": [
                      {
                        "id": 10001,
                        "orderNo": "ORD20260411123456",
                        "status": "pending",
                        "totalAmount": 299
                      }
                    ],
                    "total": 50
                  },
                  "message": "获取成功"
                }
              }
            },
            "headers": {},
            "x-apifox-name": "成功",
            "x-apifox-ordering": 0
          },
          "400": {
            "description": "请求参数错误",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                },
                "example": {
                  "code": 400,
                  "message": "请求参数错误",
                  "details": {
                    "field": "name",
                    "reason": "不能为空"
                  }
                }
              }
            },
            "headers": {},
            "x-apifox-name": "请求有误",
            "x-apifox-ordering": 1
          },
          "401": {
            "description": "未授权",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                },
                "example": {
                  "code": 401,
                  "message": "未授权，请先登录"
                }
              }
            },
            "headers": {},
            "x-apifox-name": "没有权限",
            "x-apifox-ordering": 2
          },
          "403": {
            "description": "禁止访问",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                },
                "example": {
                  "code": 403,
                  "message": "无权限访问此资源"
                }
              }
            },
            "headers": {},
            "x-apifox-name": "禁止访问",
            "x-apifox-ordering": 3
          },
          "404": {
            "description": "资源不存在",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                },
                "example": {
                  "code": 404,
                  "message": "请求的资源不存在"
                }
              }
            },
            "headers": {},
            "x-apifox-name": "记录不存在",
            "x-apifox-ordering": 4
          },
          "500": {
            "description": "服务器内部错误",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                },
                "example": {
                  "code": 500,
                  "message": "服务器内部错误，请稍后重试"
                }
              }
            },
            "headers": {},
            "x-apifox-name": "服务器错误",
            "x-apifox-ordering": 5
          },
          "502": {
            "description": "网关错误",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                },
                "example": {
                  "code": 502,
                  "message": "网关错误，上游服务不可用"
                }
              }
            },
            "headers": {},
            "x-apifox-name": "网关错误",
            "x-apifox-ordering": 6
          },
          "503": {
            "description": "服务不可用",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                },
                "example": {
                  "code": 503,
                  "message": "服务暂时不可用，请稍后重试"
                }
              }
            },
            "headers": {},
            "x-apifox-name": "服务器故障",
            "x-apifox-ordering": 7
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "x-apifox-folder": "订单模块",
        "x-apifox-status": "released",
        "x-run-in-apifox": "https://app.apifox.com/web/project/8090070/apis/api-441880946-run"
      }
    },
    "/api/v1/orders/{id}": {
      "get": {
        "summary": "获取订单详情",
        "deprecated": false,
        "description": "【版本】v1\n【环境】REST 接口\n【前置条件】需要用户登录\n【功能描述】根据订单ID获取订单详细信息及包含的商品列表",
        "operationId": "获取订单详情",
        "tags": [
          "订单管理"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "订单唯一ID",
            "required": true,
            "example": 0,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "成功",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OrdersResponse"
                },
                "example": {
                  "code": 200,
                  "data": {
                    "address": "北京市海淀区中关村南大街1号",
                    "id": 10001,
                    "items": [
                      {
                        "price": 299,
                        "productId": 2005,
                        "productName": "Lumina 显示器",
                        "quantity": 1
                      }
                    ],
                    "orderNo": "ORD20260411123456",
                    "status": "pending",
                    "totalAmount": 299
                  },
                  "message": "成功"
                }
              }
            },
            "headers": {},
            "x-apifox-name": "成功",
            "x-apifox-ordering": 0
          },
          "400": {
            "description": "请求参数错误",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                },
                "example": {
                  "code": 400,
                  "message": "请求参数错误",
                  "details": {
                    "field": "name",
                    "reason": "不能为空"
                  }
                }
              }
            },
            "headers": {},
            "x-apifox-name": "请求有误",
            "x-apifox-ordering": 1
          },
          "401": {
            "description": "未授权",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                },
                "example": {
                  "code": 401,
                  "message": "未授权，请先登录"
                }
              }
            },
            "headers": {},
            "x-apifox-name": "没有权限",
            "x-apifox-ordering": 2
          },
          "403": {
            "description": "禁止访问",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                },
                "example": {
                  "code": 403,
                  "message": "无权限访问此资源"
                }
              }
            },
            "headers": {},
            "x-apifox-name": "禁止访问",
            "x-apifox-ordering": 3
          },
          "404": {
            "description": "资源不存在",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                },
                "example": {
                  "code": 404,
                  "message": "请求的资源不存在"
                }
              }
            },
            "headers": {},
            "x-apifox-name": "记录不存在",
            "x-apifox-ordering": 4
          },
          "500": {
            "description": "服务器内部错误",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                },
                "example": {
                  "code": 500,
                  "message": "服务器内部错误，请稍后重试"
                }
              }
            },
            "headers": {},
            "x-apifox-name": "服务器错误",
            "x-apifox-ordering": 5
          },
          "502": {
            "description": "网关错误",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                },
                "example": {
                  "code": 502,
                  "message": "网关错误，上游服务不可用"
                }
              }
            },
            "headers": {},
            "x-apifox-name": "网关错误",
            "x-apifox-ordering": 6
          },
          "503": {
            "description": "服务不可用",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                },
                "example": {
                  "code": 503,
                  "message": "服务暂时不可用，请稍后重试"
                }
              }
            },
            "headers": {},
            "x-apifox-name": "服务器故障",
            "x-apifox-ordering": 7
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "x-apifox-folder": "订单模块",
        "x-apifox-status": "released",
        "x-run-in-apifox": "https://app.apifox.com/web/project/8090070/apis/api-441880948-run"
      }
    },
    "/api/v1/orders/{id}/cancel": {
      "put": {
        "summary": "取消订单",
        "deprecated": false,
        "description": "【版本】v1\n【环境】REST 接口\n【前置条件】需要用户登录\n【功能描述】取消一个处于待支付状态的订单",
        "operationId": "取消订单",
        "tags": [
          "订单管理"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "需取消的订单ID",
            "required": true,
            "example": 0,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateCancelRequest"
              },
              "example": {
                "reason": "不想买了"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "成功",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UpdateCancelResponse"
                },
                "example": {
                  "code": 200,
                  "message": "订单已取消"
                }
              }
            },
            "headers": {},
            "x-apifox-name": "成功",
            "x-apifox-ordering": 0
          },
          "400": {
            "description": "请求参数错误",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                },
                "example": {
                  "code": 400,
                  "message": "请求参数错误",
                  "details": {
                    "field": "name",
                    "reason": "不能为空"
                  }
                }
              }
            },
            "headers": {},
            "x-apifox-name": "请求有误",
            "x-apifox-ordering": 1
          },
          "401": {
            "description": "未授权",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                },
                "example": {
                  "code": 401,
                  "message": "未授权，请先登录"
                }
              }
            },
            "headers": {},
            "x-apifox-name": "没有权限",
            "x-apifox-ordering": 2
          },
          "403": {
            "description": "禁止访问",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                },
                "example": {
                  "code": 403,
                  "message": "无权限访问此资源"
                }
              }
            },
            "headers": {},
            "x-apifox-name": "禁止访问",
            "x-apifox-ordering": 3
          },
          "404": {
            "description": "资源不存在",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                },
                "example": {
                  "code": 404,
                  "message": "请求的资源不存在"
                }
              }
            },
            "headers": {},
            "x-apifox-name": "记录不存在",
            "x-apifox-ordering": 4
          },
          "409": {
            "description": "资源冲突",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                },
                "example": {
                  "code": 409,
                  "message": "资源已存在或状态冲突"
                }
              }
            },
            "headers": {},
            "x-apifox-ordering": 5
          },
          "422": {
            "description": "实体无法处理",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                },
                "example": {
                  "code": 422,
                  "message": "请求格式正确但语义错误",
                  "details": {
                    "field": "email",
                    "reason": "格式不正确"
                  }
                }
              }
            },
            "headers": {},
            "x-apifox-name": "参数错误",
            "x-apifox-ordering": 6
          },
          "500": {
            "description": "服务器内部错误",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                },
                "example": {
                  "code": 500,
                  "message": "服务器内部错误，请稍后重试"
                }
              }
            },
            "headers": {},
            "x-apifox-name": "服务器错误",
            "x-apifox-ordering": 7
          },
          "502": {
            "description": "网关错误",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                },
                "example": {
                  "code": 502,
                  "message": "网关错误，上游服务不可用"
                }
              }
            },
            "headers": {},
            "x-apifox-name": "网关错误",
            "x-apifox-ordering": 8
          },
          "503": {
            "description": "服务不可用",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                },
                "example": {
                  "code": 503,
                  "message": "服务暂时不可用，请稍后重试"
                }
              }
            },
            "headers": {},
            "x-apifox-name": "服务器故障",
            "x-apifox-ordering": 9
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "x-apifox-folder": "订单模块",
        "x-apifox-status": "released",
        "x-run-in-apifox": "https://app.apifox.com/web/project/8090070/apis/api-441880951-run"
      }
    },
...
    "/api/v1/payments/orders/{orderId}": {
      "get": {
        "summary": "根据订单ID获取支付信息",
        "deprecated": false,
        "description": "【版本】v1\n【环境】REST 接口\n【前置条件】需要用户登录\n【功能描述】根据订单ID查询对应的支付状态及详情",
        "operationId": "根据订单id获取支付信息",
        "tags": [
          "支付管理"
        ],
        "parameters": [
          {
            "name": "orderId",
            "in": "path",
            "description": "原订单的唯一ID",
            "required": true,
            "example": 0,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "成功",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OrderResponse"
                },
                "example": {
                  "code": 200,
                  "data": {
                    "amount": 299,
                    "orderId": 10001,
                    "paidAt": "2026-04-11T17:45:00Z",
                    "paymentId": "PAY20260411123456",
                    "paymentMethod": "alipay",
                    "status": "success"
                  },
                  "message": "查询成功"
                }
              }
            },
            "headers": {},
            "x-apifox-name": "成功",
            "x-apifox-ordering": 0
          },
          "400": {
            "description": "请求参数错误",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                },
                "example": {
                  "code": 400,
                  "message": "请求参数错误",
                  "details": {
                    "field": "name",
                    "reason": "不能为空"
                  }
                }
              }
            },
            "headers": {},
            "x-apifox-name": "请求有误",
            "x-apifox-ordering": 1
          },
          "401": {
            "description": "未授权",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                },
                "example": {
                  "code": 401,
                  "message": "未授权，请先登录"
                }
              }
            },
            "headers": {},
            "x-apifox-name": "没有权限",
            "x-apifox-ordering": 2
          },
          "403": {
            "description": "禁止访问",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                },
                "example": {
                  "code": 403,
                  "message": "无权限访问此资源"
                }
              }
            },
            "headers": {},
            "x-apifox-name": "禁止访问",
            "x-apifox-ordering": 3
          },
          "404": {
            "description": "资源不存在",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                },
                "example": {
                  "code": 404,
                  "message": "请求的资源不存在"
                }
              }
            },
            "headers": {},
            "x-apifox-name": "记录不存在",
            "x-apifox-ordering": 4
          },
          "500": {
            "description": "服务器内部错误",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                },
                "example": {
                  "code": 500,
                  "message": "服务器内部错误，请稍后重试"
                }
              }
            },
            "headers": {},
            "x-apifox-name": "服务器错误",
            "x-apifox-ordering": 5
          },
          "502": {
            "description": "网关错误",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                },
                "example": {
                  "code": 502,
                  "message": "网关错误，上游服务不可用"
                }
              }
            },
            "headers": {},
            "x-apifox-name": "网关错误",
            "x-apifox-ordering": 6
          },
          "503": {
            "description": "服务不可用",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                },
                "example": {
                  "code": 503,
                  "message": "服务暂时不可用，请稍后重试"
                }
              }
            },
            "headers": {},
            "x-apifox-name": "服务器故障",
            "x-apifox-ordering": 7
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "x-apifox-folder": "支付模块",
        "x-apifox-status": "released",
        "x-run-in-apifox": "https://app.apifox.com/web/project/8090070/apis/api-441881411-run"
      }
    },
...
  },
...
  "components": {
    "schemas": {
...
      "ErrorResponse": {
        "type": "object",
        "description": "标准错误响应",
        "properties": {
          "code": {
            "type": "integer",
            "description": "错误码"
          },
          "message": {
            "type": "string",
            "description": "错误信息"
          },
          "details": {
            "type": "object",
            "description": "详细信息",
            "x-apifox-orders": [],
            "properties": {},
            "x-apifox-ignore-properties": []
          }
        },
        "required": [
          "code",
          "message"
        ],
        "x-apifox-orders": [
          "code",
          "message",
          "details"
        ],
        "x-apifox-ignore-properties": [],
        "x-apifox-folder": ""
      },
...
      "CreateOrdersRequest": {
        "properties": {
          "addressId": {
            "description": "收货地址ID",
            "type": "integer"
          },
          "items": {
            "description": "订单商品列表",
            "items": {
              "properties": {
                "productId": {
                  "description": "商品ID",
                  "type": "integer"
                },
                "quantity": {
                  "description": "购买数量",
                  "type": "integer"
                }
              },
              "type": "object",
              "x-apifox-orders": [
                "productId",
                "quantity"
              ],
              "x-apifox-ignore-properties": []
            },
            "type": "array"
          },
          "note": {
            "description": "订单备注",
            "type": "string"
          }
        },
        "required": [
          "addressId"
        ],
        "type": "object",
        "x-apifox-orders": [
          "addressId",
          "items",
          "note"
        ],
        "x-apifox-ignore-properties": [],
        "x-apifox-folder": ""
      },
...
      "UpdateCancelRequest": {
        "properties": {
          "reason": {
            "description": "取消原因",
            "type": "string"
          }
        },
        "type": "object",
        "x-apifox-orders": [
          "reason"
        ],
        "x-apifox-ignore-properties": [],
        "x-apifox-folder": ""
      },
...
      "CreateOrdersResponse": {
        "properties": {
          "code": {
            "description": "状态码",
            "type": "integer"
          },
          "data": {
            "description": "订单数据",
            "properties": {
              "orderId": {
                "description": "新建订单的唯一ID",
                "type": "integer"
              },
              "orderNo": {
                "description": "订单编号",
                "type": "string"
              }
            },
            "type": "object",
            "x-apifox-orders": [
              "orderId",
              "orderNo"
            ],
            "x-apifox-ignore-properties": []
          },
          "message": {
            "description": "提示信息",
            "type": "string"
          }
        },
        "required": [
          "code",
          "message"
        ],
        "type": "object",
        "x-apifox-orders": [
          "code",
          "data",
          "message"
        ],
        "x-apifox-ignore-properties": [],
        "x-apifox-folder": ""
      },
...
    },
...
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  },
  "servers": [],
  "security": []
}